refactor: Separate source file of each primitive type
This makes it easy to focus on one type at a time and reduces the size of already long resolved_builder.rs and resolved_type.rs
