+++ crates/postgres-subsystem/postgres-core-builder/src/resolved_builder.rs
@@ -13,9 +13,6 @@
 //! column name, here that information is encoded into an attribute of `ResolvedType`.
 //! If no @column is provided, the encoded information is set to an appropriate default value.
 
-use std::collections::{HashMap, HashSet};
-use std::sync::LazyLock;
-
 use codemap_diagnostic::{Diagnostic, Level, SpanLabel, SpanStyle};
 use postgres_core_model::types::EntityRepresentation;
 use serde::{Deserialize, Serialize};
@@ -24,15 +21,16 @@ use super::{
     access_builder::{ResolvedAccess, build_access},
     naming::{ToPlural, ToTableName},
 };
-use crate::resolved_type::{
-    DateTimeTypeHint, DecimalTypeHint, ExplicitTypeHint, FloatTypeHint, IntTypeHint,
-    ResolvedCompositeType, ResolvedEnumType, ResolvedField, ResolvedFieldDefault,
-    ResolvedFieldType, ResolvedType, ResolvedTypeHint, SerializableTypeHint, StringTypeHint,
-    VectorTypeHint,
+use crate::{
+    resolved_type::{
+        ExplicitTypeHint, ResolvedCompositeType, ResolvedEnumType, ResolvedField,
+        ResolvedFieldDefault, ResolvedFieldType, ResolvedType, SerializableTypeHint,
+    },
+    type_provider::{PRIMITIVE_TYPE_PROVIDER_REGISTRY, validate_hint_annotations},
 };
 use core_model::{
     mapped_arena::MappedArena,
-    primitive_type::{self, PrimitiveBaseType},
+    primitive_type::{self},
     types::{FieldType, Named},
 };
 use core_model_builder::{
@@ -44,1048 +42,16 @@ use core_model_builder::{
     error::ModelBuildingError,
     typechecker::{
         Typed,
-        annotation::{AnnotationSpec, AnnotationTarget, MappedAnnotationParamSpec},
         typ::{Module, Type, TypecheckedSystem},
     },
 };
-use exo_sql::{
-    DEFAULT_VECTOR_SIZE, FloatBits, IntBits, PhysicalColumnType, SchemaObjectName,
-    VectorDistanceFunction,
-};
-use postgres_core_model::aggregate::ScalarAggregateFieldKind;
+use exo_sql::SchemaObjectName;
 
 use heck::ToSnakeCase;
 
 const DEFAULT_FN_AUTO_INCREMENT: &str = "autoIncrement";
 const DEFAULT_FN_CURRENT_TIME: &str = "now";
 const DEFAULT_FN_GENERATE_UUID: &str = "generate_uuid";
-
-/// Provide Postgres-specific functionality for primitive types
-/// Handles physical column type determination, type hints, annotations, and deserialization
-pub trait PrimitiveTypeProvider: Send + Sync + PrimitiveBaseType {
-    /// Determines the physical column type for a field with this primitive type
-    fn determine_column_type(&self, field: &ResolvedField) -> PhysicalColumnType;
-
-    /// Computes the type hint for a field, validating that only supported hint annotations are used.
-    fn compute_type_hint(
-        &self,
-        field: &AstField<Typed>,
-        errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint>;
-
-    /// Deserialize JSON data into a type hint for this provider
-    /// Returns an error for primitive types that don't support type hints.
-    fn deserialize_type_hint(
-        &self,
-        _data: serde_json::Value,
-    ) -> Result<Box<dyn ResolvedTypeHint>, String> {
-        Err(format!("Type {} does not support type hints", self.name()))
-    }
-
-    /// Get applicable hint annotations for this type
-    /// Returns an empty vector for types that don't support type hints.
-    fn applicable_hint_annotations(&self) -> Vec<(&'static str, AnnotationSpec)> {
-        vec![]
-    }
-
-    /// Get supported operators for this primitive type
-    /// Returns None if the type supports no operators (implicit equality only)
-    /// Returns Some(operators) if the type supports specific operators
-    fn supported_operators(&self) -> Option<Vec<&'static str>>;
-
-    /// Get supported aggregate functions for this primitive type
-    /// Returns a vector of (aggregate_kind, optional_return_type) tuples
-    /// The return type is Some when the aggregate function returns a different type than the input
-    /// (e.g., avg of Int returns Float). The "count" aggregate is always supported and doesn't need to be listed here.
-    fn supported_aggregates(
-        &self,
-    ) -> Vec<(
-        ScalarAggregateFieldKind,
-        Option<&'static dyn PrimitiveBaseType>,
-    )> {
-        vec![]
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::IntType {
-    fn determine_column_type(&self, field: &ResolvedField) -> PhysicalColumnType {
-        match &field.type_hint {
-            Some(hint) => {
-                let hint_ref = hint.0.as_ref() as &dyn std::any::Any;
-
-                if let Some(int_hint) = hint_ref.downcast_ref::<IntTypeHint>() {
-                    // determine the proper sized type to use
-                    if let Some(bits) = int_hint.bits {
-                        PhysicalColumnType::Int {
-                            bits: match bits {
-                                16 => IntBits::_16,
-                                32 => IntBits::_32,
-                                64 => IntBits::_64,
-                                _ => panic!("Invalid bits"),
-                            },
-                        }
-                    } else if let Some(range) = &int_hint.range {
-                        let is_superset = |bound_min: i64, bound_max: i64| {
-                            let range_min = range.0;
-                            let range_max = range.1;
-                            assert!(range_min <= range_max);
-                            assert!(bound_min <= bound_max);
-
-                            // is this bound a superset of the provided range?
-                            (bound_min <= range_min && bound_min <= range_max)
-                                && (bound_max >= range_max && bound_max >= range_min)
-                        };
-
-                        // determine which SQL type is appropriate for this range
-                        if is_superset(i16::MIN.into(), i16::MAX.into()) {
-                            PhysicalColumnType::Int { bits: IntBits::_16 }
-                        } else if is_superset(i32::MIN.into(), i32::MAX.into()) {
-                            PhysicalColumnType::Int { bits: IntBits::_32 }
-                        } else if is_superset(i64::MIN, i64::MAX) {
-                            PhysicalColumnType::Int { bits: IntBits::_64 }
-                        } else {
-                            // TODO: numeric type
-                            panic!("Requested range is too big")
-                        }
-                    } else {
-                        // no specific hints provided, go with default
-                        PhysicalColumnType::Int { bits: IntBits::_32 }
-                    }
-                } else {
-                    // no relevant hints provided, go with default
-                    PhysicalColumnType::Int { bits: IntBits::_32 }
-                }
-            }
-            None => {
-                // no hints provided, go with default
-                PhysicalColumnType::Int { bits: IntBits::_32 }
-            }
-        }
-    }
-
-    fn compute_type_hint(
-        &self,
-        field: &AstField<Typed>,
-        errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        let range_hint = field.annotations.get("range").map(|params| {
-            (
-                params.as_map().get("min").unwrap().as_int(),
-                params.as_map().get("max").unwrap().as_int(),
-            )
-        });
-
-        let is_bits16 = field.annotations.contains("bits16");
-        let is_bits32 = field.annotations.contains("bits32");
-        let is_bits64 = field.annotations.contains("bits64");
-
-        let bits_hint = match (is_bits16, is_bits32, is_bits64) {
-            (true, false, false) => Some(16),
-            (false, true, false) => Some(32),
-            (false, false, true) => Some(64),
-            (false, false, false) => None,
-            _ => {
-                errors.push(Diagnostic {
-                    level: Level::Error,
-                    message: "Cannot have more than one of @bits16, @bits32, @bits64".to_string(),
-                    code: Some("C000".to_string()),
-                    spans: vec![SpanLabel {
-                        span: field.span,
-                        style: SpanStyle::Primary,
-                        label: None,
-                    }],
-                });
-                None
-            }
-        };
-
-        if bits_hint.is_some() || range_hint.is_some() {
-            Some(SerializableTypeHint(Box::new(IntTypeHint {
-                bits: bits_hint,
-                range: range_hint,
-            })))
-        } else {
-            None
-        }
-    }
-
-    fn deserialize_type_hint(
-        &self,
-        data: serde_json::Value,
-    ) -> Result<Box<dyn ResolvedTypeHint>, String> {
-        let hint: IntTypeHint = serde_json::from_value(data)
-            .map_err(|e| format!("Failed to deserialize IntTypeHint: {}", e))?;
-        Ok(Box::new(hint))
-    }
-
-    fn applicable_hint_annotations(&self) -> Vec<(&'static str, AnnotationSpec)> {
-        vec![
-            (
-                "range",
-                AnnotationSpec {
-                    targets: &[AnnotationTarget::Field],
-                    no_params: false,
-                    single_params: false,
-                    mapped_params: Some(&[
-                        MappedAnnotationParamSpec {
-                            name: "min",
-                            optional: false,
-                        },
-                        MappedAnnotationParamSpec {
-                            name: "max",
-                            optional: false,
-                        },
-                    ]),
-                },
-            ),
-            (
-                "bits16",
-                AnnotationSpec {
-                    targets: &[AnnotationTarget::Field],
-                    no_params: true,
-                    single_params: false,
-                    mapped_params: None,
-                },
-            ),
-            (
-                "bits32",
-                AnnotationSpec {
-                    targets: &[AnnotationTarget::Field],
-                    no_params: true,
-                    single_params: false,
-                    mapped_params: None,
-                },
-            ),
-            (
-                "bits64",
-                AnnotationSpec {
-                    targets: &[AnnotationTarget::Field],
-                    no_params: true,
-                    single_params: false,
-                    mapped_params: None,
-                },
-            ),
-        ]
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec!["eq", "neq", "lt", "lte", "gt", "gte"])
-    }
-
-    fn supported_aggregates(
-        &self,
-    ) -> Vec<(
-        ScalarAggregateFieldKind,
-        Option<&'static dyn PrimitiveBaseType>,
-    )> {
-        use ScalarAggregateFieldKind::*;
-        vec![
-            (Min, None),
-            (Max, None),
-            (Sum, None),
-            (Avg, Some(&primitive_type::FloatType)),
-        ]
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::FloatType {
-    fn determine_column_type(&self, field: &ResolvedField) -> PhysicalColumnType {
-        match &field.type_hint {
-            Some(hint) => {
-                let hint_ref = hint.0.as_ref() as &dyn std::any::Any;
-
-                if let Some(float_hint) = hint_ref.downcast_ref::<FloatTypeHint>() {
-                    if let Some(bits) = float_hint.bits {
-                        if (1..=24).contains(&bits) {
-                            PhysicalColumnType::Float {
-                                bits: FloatBits::_24,
-                            }
-                        } else if bits > 24 && bits <= 53 {
-                            PhysicalColumnType::Float {
-                                bits: FloatBits::_53,
-                            }
-                        } else {
-                            panic!("Invalid bits")
-                        }
-                    } else {
-                        PhysicalColumnType::Float {
-                            bits: FloatBits::_53,
-                        }
-                    }
-                } else {
-                    PhysicalColumnType::Float {
-                        bits: FloatBits::_24,
-                    }
-                }
-            }
-            None => PhysicalColumnType::Float {
-                bits: FloatBits::_24,
-            },
-        }
-    }
-
-    fn compute_type_hint(
-        &self,
-        field: &AstField<Typed>,
-        errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        let mut range_hint = None;
-        if let Some(params) = field.annotations.get("range") {
-            let min = params
-                .as_map()
-                .get("min")
-                .unwrap()
-                .as_string()
-                .parse::<f64>();
-            let max = params
-                .as_map()
-                .get("max")
-                .unwrap()
-                .as_string()
-                .parse::<f64>();
-
-            if min.is_err() || max.is_err() {
-                if min.is_err() {
-                    errors.push(Diagnostic {
-                        level: Level::Error,
-                        message: "Cannot parse @range 'min' as f64".to_string(),
-                        code: Some("C000".to_string()),
-                        spans: vec![SpanLabel {
-                            span: field.span,
-                            style: SpanStyle::Primary,
-                            label: None,
-                        }],
-                    });
-                }
-                if max.is_err() {
-                    errors.push(Diagnostic {
-                        level: Level::Error,
-                        message: "Cannot parse @range 'max' as f64".to_string(),
-                        code: Some("C000".to_string()),
-                        spans: vec![SpanLabel {
-                            span: field.span,
-                            style: SpanStyle::Primary,
-                            label: None,
-                        }],
-                    });
-                }
-            } else {
-                range_hint = Some((min.unwrap(), max.unwrap()));
-            }
-        };
-
-        let is_single_precision = field.annotations.contains("singlePrecision");
-        let is_double_precision = field.annotations.contains("doublePrecision");
-
-        let bits_hint = match (is_single_precision, is_double_precision) {
-            (true, false) => Some(24),
-            (false, true) => Some(53),
-            (false, false) => None,
-            _ => {
-                errors.push(Diagnostic {
-                    level: Level::Error,
-                    message: "Cannot have both @singlePrecision and @doublePrecision".to_string(),
-                    code: Some("C000".to_string()),
-                    spans: vec![SpanLabel {
-                        span: field.span,
-                        style: SpanStyle::Primary,
-                        label: None,
-                    }],
-                });
-                None
-            }
-        };
-
-        if bits_hint.is_some() || range_hint.is_some() {
-            Some(SerializableTypeHint(Box::new(FloatTypeHint {
-                bits: bits_hint,
-                range: range_hint,
-            })))
-        } else {
-            None
-        }
-    }
-
-    fn deserialize_type_hint(
-        &self,
-        data: serde_json::Value,
-    ) -> Result<Box<dyn ResolvedTypeHint>, String> {
-        let hint: FloatTypeHint = serde_json::from_value(data)
-            .map_err(|e| format!("Failed to deserialize FloatTypeHint: {}", e))?;
-        Ok(Box::new(hint))
-    }
-
-    fn applicable_hint_annotations(&self) -> Vec<(&'static str, AnnotationSpec)> {
-        vec![
-            (
-                "range",
-                AnnotationSpec {
-                    targets: &[AnnotationTarget::Field],
-                    no_params: false,
-                    single_params: false,
-                    mapped_params: Some(&[
-                        MappedAnnotationParamSpec {
-                            name: "min",
-                            optional: false,
-                        },
-                        MappedAnnotationParamSpec {
-                            name: "max",
-                            optional: false,
-                        },
-                    ]),
-                },
-            ),
-            (
-                "singlePrecision",
-                AnnotationSpec {
-                    targets: &[AnnotationTarget::Field],
-                    no_params: true,
-                    single_params: false,
-                    mapped_params: None,
-                },
-            ),
-            (
-                "doublePrecision",
-                AnnotationSpec {
-                    targets: &[AnnotationTarget::Field],
-                    no_params: true,
-                    single_params: false,
-                    mapped_params: None,
-                },
-            ),
-        ]
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec!["eq", "neq", "lt", "lte", "gt", "gte"])
-    }
-
-    fn supported_aggregates(
-        &self,
-    ) -> Vec<(
-        ScalarAggregateFieldKind,
-        Option<&'static dyn PrimitiveBaseType>,
-    )> {
-        use ScalarAggregateFieldKind::*;
-        vec![
-            (Min, None),
-            (Max, None),
-            (Sum, Some(&primitive_type::FloatType)),
-            (Avg, Some(&primitive_type::FloatType)),
-        ]
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::DecimalType {
-    fn determine_column_type(&self, field: &ResolvedField) -> PhysicalColumnType {
-        match &field.type_hint {
-            Some(hint) => {
-                let hint_ref = hint.0.as_ref() as &dyn std::any::Any;
-
-                if let Some(decimal_hint) = hint_ref.downcast_ref::<DecimalTypeHint>() {
-                    // cannot have scale and no precision specified
-                    if decimal_hint.precision.is_none() {
-                        assert!(decimal_hint.scale.is_none())
-                    }
-
-                    PhysicalColumnType::Numeric {
-                        precision: decimal_hint.precision,
-                        scale: decimal_hint.scale,
-                    }
-                } else {
-                    PhysicalColumnType::Numeric {
-                        precision: None,
-                        scale: None,
-                    }
-                }
-            }
-            None => PhysicalColumnType::Numeric {
-                precision: None,
-                scale: None,
-            },
-        }
-    }
-
-    fn compute_type_hint(
-        &self,
-        field: &AstField<Typed>,
-        errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        let precision_hint = field
-            .annotations
-            .get("precision")
-            .map(|p| p.as_single().as_int() as usize);
-
-        let scale_hint = field
-            .annotations
-            .get("scale")
-            .map(|p| p.as_single().as_int() as usize);
-
-        if scale_hint.is_some() && precision_hint.is_none() {
-            errors.push(Diagnostic {
-                level: Level::Error,
-                message: "@scale is not allowed without specifying @precision".to_string(),
-                code: Some("C000".to_string()),
-                spans: vec![SpanLabel {
-                    span: field.span,
-                    style: SpanStyle::Primary,
-                    label: None,
-                }],
-            });
-        }
-
-        Some(SerializableTypeHint(Box::new(DecimalTypeHint {
-            precision: precision_hint,
-            scale: scale_hint,
-        })))
-    }
-
-    fn deserialize_type_hint(
-        &self,
-        data: serde_json::Value,
-    ) -> Result<Box<dyn ResolvedTypeHint>, String> {
-        let hint: DecimalTypeHint = serde_json::from_value(data)
-            .map_err(|e| format!("Failed to deserialize DecimalTypeHint: {}", e))?;
-        Ok(Box::new(hint))
-    }
-
-    fn applicable_hint_annotations(&self) -> Vec<(&'static str, AnnotationSpec)> {
-        vec![
-            (
-                "precision",
-                AnnotationSpec {
-                    targets: &[AnnotationTarget::Field],
-                    no_params: false,
-                    single_params: true,
-                    mapped_params: None,
-                },
-            ),
-            (
-                "scale",
-                AnnotationSpec {
-                    targets: &[AnnotationTarget::Field],
-                    no_params: false,
-                    single_params: true,
-                    mapped_params: None,
-                },
-            ),
-        ]
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec!["eq", "neq", "lt", "lte", "gt", "gte"])
-    }
-
-    fn supported_aggregates(
-        &self,
-    ) -> Vec<(
-        ScalarAggregateFieldKind,
-        Option<&'static dyn PrimitiveBaseType>,
-    )> {
-        use ScalarAggregateFieldKind::*;
-        vec![(Min, None), (Max, None), (Sum, None), (Avg, None)]
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::StringType {
-    fn determine_column_type(&self, field: &ResolvedField) -> PhysicalColumnType {
-        match &field.type_hint {
-            Some(hint) => {
-                let hint_ref = hint.0.as_ref() as &dyn std::any::Any;
-
-                if let Some(string_hint) = hint_ref.downcast_ref::<StringTypeHint>() {
-                    // length hint provided, use it
-                    PhysicalColumnType::String {
-                        max_length: Some(string_hint.max_length),
-                    }
-                } else {
-                    PhysicalColumnType::String { max_length: None }
-                }
-            }
-            None => PhysicalColumnType::String { max_length: None },
-        }
-    }
-
-    fn compute_type_hint(
-        &self,
-        field: &AstField<Typed>,
-        _errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        let max_length_annotation = field
-            .annotations
-            .get("maxLength")
-            .map(|p| p.as_single().as_int() as usize);
-
-        max_length_annotation
-            .map(|max_length| SerializableTypeHint(Box::new(StringTypeHint { max_length })))
-    }
-
-    fn deserialize_type_hint(
-        &self,
-        data: serde_json::Value,
-    ) -> Result<Box<dyn ResolvedTypeHint>, String> {
-        let hint: StringTypeHint = serde_json::from_value(data)
-            .map_err(|e| format!("Failed to deserialize StringTypeHint: {}", e))?;
-        Ok(Box::new(hint))
-    }
-
-    fn applicable_hint_annotations(&self) -> Vec<(&'static str, AnnotationSpec)> {
-        vec![(
-            "maxLength",
-            AnnotationSpec {
-                targets: &[AnnotationTarget::Field],
-                no_params: false,
-                single_params: true,
-                mapped_params: None,
-            },
-        )]
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec![
-            "eq",
-            "neq",
-            "lt",
-            "lte",
-            "gt",
-            "gte",
-            "like",
-            "ilike",
-            "startsWith",
-            "endsWith",
-        ])
-    }
-
-    fn supported_aggregates(
-        &self,
-    ) -> Vec<(
-        ScalarAggregateFieldKind,
-        Option<&'static dyn PrimitiveBaseType>,
-    )> {
-        use ScalarAggregateFieldKind::*;
-        vec![(Min, None), (Max, None)]
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::InstantType {
-    fn determine_column_type(&self, field: &ResolvedField) -> PhysicalColumnType {
-        match &field.type_hint {
-            Some(hint) => {
-                let hint_ref = hint.0.as_ref() as &dyn std::any::Any;
-
-                if let Some(datetime_hint) = hint_ref.downcast_ref::<DateTimeTypeHint>() {
-                    PhysicalColumnType::Timestamp {
-                        precision: Some(datetime_hint.precision),
-                        timezone: true,
-                    }
-                } else {
-                    PhysicalColumnType::Timestamp {
-                        precision: None,
-                        timezone: true,
-                    }
-                }
-            }
-            None => PhysicalColumnType::Timestamp {
-                precision: None,
-                timezone: true,
-            },
-        }
-    }
-
-    fn compute_type_hint(
-        &self,
-        field: &AstField<Typed>,
-        _errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        field.annotations.get("precision").map(|p| {
-            SerializableTypeHint(Box::new(DateTimeTypeHint {
-                precision: p.as_single().as_int() as usize,
-            }))
-        })
-    }
-
-    fn deserialize_type_hint(
-        &self,
-        data: serde_json::Value,
-    ) -> Result<Box<dyn ResolvedTypeHint>, String> {
-        let hint: DateTimeTypeHint = serde_json::from_value(data)
-            .map_err(|e| format!("Failed to deserialize DateTimeTypeHint: {}", e))?;
-        Ok(Box::new(hint))
-    }
-
-    fn applicable_hint_annotations(&self) -> Vec<(&'static str, AnnotationSpec)> {
-        vec![(
-            "precision",
-            AnnotationSpec {
-                targets: &[AnnotationTarget::Field],
-                no_params: false,
-                single_params: true,
-                mapped_params: None,
-            },
-        )]
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec!["eq", "neq", "lt", "lte", "gt", "gte"])
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::VectorType {
-    fn determine_column_type(&self, field: &ResolvedField) -> PhysicalColumnType {
-        match &field.type_hint {
-            Some(hint) => {
-                let hint_ref = hint.0.as_ref() as &dyn std::any::Any;
-
-                if let Some(vector_hint) = hint_ref.downcast_ref::<VectorTypeHint>() {
-                    PhysicalColumnType::Vector {
-                        size: vector_hint.size.unwrap_or(DEFAULT_VECTOR_SIZE),
-                    }
-                } else {
-                    PhysicalColumnType::Vector {
-                        size: DEFAULT_VECTOR_SIZE,
-                    }
-                }
-            }
-            None => PhysicalColumnType::Vector {
-                size: DEFAULT_VECTOR_SIZE,
-            },
-        }
-    }
-
-    fn compute_type_hint(
-        &self,
-        field: &AstField<Typed>,
-        errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        let size = field
-            .annotations
-            .get("size")
-            .map(|p| p.as_single().as_int() as usize);
-
-        let distance_function = field.annotations.get("distanceFunction").and_then(|p| {
-            match VectorDistanceFunction::from_model_string(p.as_single().as_string().as_str()) {
-                Ok(distance_function) => Some(distance_function),
-                Err(e) => {
-                    errors.push(Diagnostic {
-                        level: Level::Error,
-                        message: e.to_string(),
-                        code: Some("C000".to_string()),
-                        spans: vec![SpanLabel {
-                            span: field.span,
-                            style: SpanStyle::Primary,
-                            label: None,
-                        }],
-                    });
-                    None
-                }
-            }
-        });
-
-        Some(SerializableTypeHint(Box::new(VectorTypeHint {
-            size,
-            distance_function,
-        })))
-    }
-
-    fn deserialize_type_hint(
-        &self,
-        data: serde_json::Value,
-    ) -> Result<Box<dyn ResolvedTypeHint>, String> {
-        let hint: VectorTypeHint = serde_json::from_value(data)
-            .map_err(|e| format!("Failed to deserialize VectorTypeHint: {}", e))?;
-        Ok(Box::new(hint))
-    }
-
-    fn applicable_hint_annotations(&self) -> Vec<(&'static str, AnnotationSpec)> {
-        vec![
-            (
-                "size",
-                AnnotationSpec {
-                    targets: &[AnnotationTarget::Field],
-                    no_params: false,
-                    single_params: true,
-                    mapped_params: None,
-                },
-            ),
-            (
-                "distanceFunction",
-                AnnotationSpec {
-                    targets: &[AnnotationTarget::Field],
-                    no_params: false,
-                    single_params: true,
-                    mapped_params: None,
-                },
-            ),
-        ]
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec!["similar", "eq", "neq"])
-    }
-
-    fn supported_aggregates(
-        &self,
-    ) -> Vec<(
-        ScalarAggregateFieldKind,
-        Option<&'static dyn PrimitiveBaseType>,
-    )> {
-        use ScalarAggregateFieldKind::*;
-        vec![(Avg, None)]
-    }
-}
-
-// Special implementations for time types that can have datetime hints
-impl PrimitiveTypeProvider for primitive_type::LocalTimeType {
-    fn determine_column_type(&self, field: &ResolvedField) -> PhysicalColumnType {
-        match &field.type_hint {
-            Some(hint) => {
-                let hint_ref = hint.0.as_ref() as &dyn std::any::Any;
-
-                if let Some(datetime_hint) = hint_ref.downcast_ref::<DateTimeTypeHint>() {
-                    PhysicalColumnType::Time {
-                        precision: Some(datetime_hint.precision),
-                    }
-                } else {
-                    PhysicalColumnType::Time { precision: None }
-                }
-            }
-            None => PhysicalColumnType::Time { precision: None },
-        }
-    }
-
-    fn compute_type_hint(
-        &self,
-        _field: &AstField<Typed>,
-        _errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        None
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec!["eq", "neq", "lt", "lte", "gt", "gte"])
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::LocalDateTimeType {
-    fn determine_column_type(&self, field: &ResolvedField) -> PhysicalColumnType {
-        match &field.type_hint {
-            Some(hint) => {
-                let hint_ref = hint.0.as_ref() as &dyn std::any::Any;
-
-                if let Some(datetime_hint) = hint_ref.downcast_ref::<DateTimeTypeHint>() {
-                    PhysicalColumnType::Timestamp {
-                        precision: Some(datetime_hint.precision),
-                        timezone: false,
-                    }
-                } else {
-                    PhysicalColumnType::Timestamp {
-                        precision: None,
-                        timezone: false,
-                    }
-                }
-            }
-            None => PhysicalColumnType::Timestamp {
-                precision: None,
-                timezone: false,
-            },
-        }
-    }
-
-    fn compute_type_hint(
-        &self,
-        _field: &AstField<Typed>,
-        _errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        None
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec!["eq", "neq", "lt", "lte", "gt", "gte"])
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::BooleanType {
-    fn determine_column_type(&self, _field: &ResolvedField) -> PhysicalColumnType {
-        PhysicalColumnType::Boolean
-    }
-
-    fn compute_type_hint(
-        &self,
-        _field: &AstField<Typed>,
-        _errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        None
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec!["eq", "neq"])
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::LocalDateType {
-    fn determine_column_type(&self, _field: &ResolvedField) -> PhysicalColumnType {
-        PhysicalColumnType::Date
-    }
-
-    fn compute_type_hint(
-        &self,
-        _field: &AstField<Typed>,
-        _errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        None
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec!["eq", "neq", "lt", "lte", "gt", "gte"])
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::JsonType {
-    fn determine_column_type(&self, _field: &ResolvedField) -> PhysicalColumnType {
-        PhysicalColumnType::Json
-    }
-
-    fn compute_type_hint(
-        &self,
-        _field: &AstField<Typed>,
-        _errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        None
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec![
-            "contains",
-            "containedBy",
-            "matchKey",
-            "matchAllKeys",
-            "matchAnyKey",
-        ])
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::BlobType {
-    fn determine_column_type(&self, _field: &ResolvedField) -> PhysicalColumnType {
-        PhysicalColumnType::Blob
-    }
-
-    fn compute_type_hint(
-        &self,
-        _field: &AstField<Typed>,
-        _errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        None
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        None
-    }
-}
-
-impl PrimitiveTypeProvider for primitive_type::UuidType {
-    fn determine_column_type(&self, _field: &ResolvedField) -> PhysicalColumnType {
-        PhysicalColumnType::Uuid
-    }
-
-    fn compute_type_hint(
-        &self,
-        _field: &AstField<Typed>,
-        _errors: &mut Vec<Diagnostic>,
-    ) -> Option<SerializableTypeHint> {
-        None
-    }
-
-    fn supported_operators(&self) -> Option<Vec<&'static str>> {
-        Some(vec!["eq", "neq"])
-    }
-}
-
-/// Unified registry mapping primitive type names to their providers
-/// Handles all primitive type functionality: column types, type hints, annotations, deserialization
-pub static PRIMITIVE_TYPE_PROVIDER_REGISTRY: LazyLock<
-    HashMap<&'static str, &'static dyn PrimitiveTypeProvider>,
-> = LazyLock::new(|| {
-    let all_primitive_type_providers: &[&dyn PrimitiveTypeProvider] = &[
-        &primitive_type::IntType,
-        &primitive_type::FloatType,
-        &primitive_type::DecimalType,
-        &primitive_type::StringType,
-        &primitive_type::InstantType,
-        &primitive_type::VectorType,
-        &primitive_type::BooleanType,
-        &primitive_type::LocalDateType,
-        &primitive_type::LocalTimeType,
-        &primitive_type::LocalDateTimeType,
-        &primitive_type::JsonType,
-        &primitive_type::BlobType,
-        &primitive_type::UuidType,
-    ];
-    all_primitive_type_providers
-        .iter()
-        .map(|provider| (provider.name(), *provider))
-        .collect()
-});
-
-static ALL_HINT_ANNOTATION_NAMES: LazyLock<HashSet<&'static str>> = LazyLock::new(|| {
-    collect_all_hint_annotations()
-        .iter()
-        .map(|(name, _)| *name)
-        .collect()
-});
-
-/// Validates that only supported hint annotations are used for a given field and type provider
-fn validate_hint_annotations(
-    field: &AstField<Typed>,
-    provider: &dyn PrimitiveTypeProvider,
-    errors: &mut Vec<Diagnostic>,
-) {
-    let field_annotations = field.annotations.annotations.keys();
-    let supported_annotations: Vec<&'static str> = provider
-        .applicable_hint_annotations()
-        .iter()
-        .map(|(name, _)| *name)
-        .collect();
-    let supported_hint_annotations_set: std::collections::HashSet<&str> =
-        supported_annotations.iter().copied().collect();
-
-    let unsupported_hint_annotations = field_annotations
-        .filter(|annotation| {
-            // Only validate against hint annotations that this type supports
-            // Skip annotations that are not hint annotations (they'll be validated elsewhere)
-            ALL_HINT_ANNOTATION_NAMES.contains(annotation.as_str())
-                && !supported_hint_annotations_set.contains(annotation.as_str())
-        })
-        .collect::<Vec<_>>();
-
-    if !unsupported_hint_annotations.is_empty() {
-        errors.push(Diagnostic {
-            level: Level::Error,
-            message: format!(
-                "Annotation @{} is not supported for type {}",
-                unsupported_hint_annotations
-                    .iter()
-                    .map(|a| a.as_str())
-                    .collect::<Vec<_>>()
-                    .join(", "),
-                provider.name()
-            ),
-            code: Some("C000".to_string()),
-            spans: vec![SpanLabel {
-                span: field.span,
-                style: SpanStyle::Primary,
-                label: None,
-            }],
-        });
-    }
-}
-
-/// Collects all hint annotations from all registered type hint providers
-pub fn collect_all_hint_annotations() -> Vec<(&'static str, AnnotationSpec)> {
-    PRIMITIVE_TYPE_PROVIDER_REGISTRY
-        .iter()
-        .flat_map(|(_, provider)| provider.applicable_hint_annotations())
-        .collect()
-}
 
 /// Consume typed-checked types and build resolved types
 pub fn build(
